requirements.txt
allure-pytest
assertpy
numpy
pandas
pytest
python-dotenv
